fix(runtime): harden retained run ownership
